Add support for 64-bit Windows build
BUG=28176
TEST=none

Review URL: http://codereview.chromium.org/455036

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@33645 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 11e1a2a71fa01d356f534989265f366006fb4f52

+++ base/event_trace_provider_win.cc
@@ -99,7 +99,7 @@ ULONG EtwTraceProvider::Log(const EtwEventClass& event_class,
 
   event.fields[0].DataPtr = reinterpret_cast<ULONG64>(message);
   event.fields[0].Length = message ?
-      sizeof(message[0]) * (1 + strlen(message)) : 0;
+      static_cast<ULONG>(sizeof(message[0]) * (1 + strlen(message))) : 0;
 
   return ::TraceEvent(session_handle_, &event.header);
 }
@@ -113,7 +113,7 @@ ULONG EtwTraceProvider::Log(const EtwEventClass& event_class,
 
   event.fields[0].DataPtr = reinterpret_cast<ULONG64>(message);
   event.fields[0].Length = message ?
-      sizeof(message[0]) * (1 + wcslen(message)) : 0;
+      static_cast<ULONG>(sizeof(message[0]) * (1 + wcslen(message))) : 0;
 
   return ::TraceEvent(session_handle_, &event.header);
 }
